Rework compose page to share the widget with quick reply
